refactor: remove interface from inside vue component
