+++ package.json
@@ -82,7 +82,7 @@
 		"start-server-and-test": "^1.14.0",
 		"storybook": "7.5.3",
 		"storybook-design-token": "3.0.0-beta.7",
-		"typescript": "~5.1.6",
+		"typescript": "^5.1.6",
 		"vite": "4.1.4",
 		"vitest": "^0.34.2",
 		"vue": "^3.2.47",
