+++ src/components/admin/page-helper-video/index.ts
@@ -1,0 +1,2 @@
+export { default as PageHelperVideo } from './PageHelperVideo.vue'
+export { type IVideo } from './types'
