+++ src/components/admin/page-helper-video/types.ts
@@ -1,0 +1,8 @@
+export interface IVideo {
+	video_id: string
+	name?: string
+	articles?: {
+		name: string
+		url: string
+	}[]
+}
